requirements.txt
blinker==1.9.0
boto3==1.35.91
botocore==1.35.91
certifi==2024.12.14
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.41.0
Deprecated==1.2.15
durationpy==0.9
Flask==3.1.0
Flask-Cors==5.0.0
hjson==3.1.0
idna==3.10
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kappa==0.6.0
lxml==5.3.0
MarkupSafe==3.0.2
packaging==24.2
pikepdf==9.5.1
pillow==11.1.0
placebo==0.9.0
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.8.3
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
zappa==0.59.0
argcomplete==3.5.3
